+++ src/components/recordHasIdAndProperties.js
