+++ test/components/recordHasIdAndPropertiesTest.js
@@ -1,7 +1,7 @@
 var tape = require('tape');
 var event_stream = require('event-stream');
 
-var hasIdAndProperties = require('../../src/components/hasIdAndProperties');
+var recordHasIdAndProperties = require('../../src/components/recordHasIdAndProperties');
 
 function test_stream(input, testedStream, callback) {
     var input_stream = event_stream.readArray(input);
@@ -10,8 +10,8 @@ function test_stream(input, testedStream, callback) {
     input_stream.pipe(testedStream).pipe(destination_stream);
 }
 
-tape('hasIdAndProperties', function(test) {
-  test.test('hasIdAndProperties filters objects without id and properties', function(t) {
+tape('recordHasIdAndProperties', function(test) {
+  test.test('recordHasIdAndProperties filters objects without id and properties', function(t) {
     var input = [
       {
         id: 5,
@@ -32,7 +32,7 @@ tape('hasIdAndProperties', function(test) {
       }
     ];
 
-    test_stream(input, hasIdAndProperties.create(), function(err, actual) {
+    test_stream(input, recordHasIdAndProperties.create(), function(err, actual) {
       t.deepEqual(actual, expected, 'stream should contain only objects with id and properties');
       t.end();
     });
