+++ test/test.js
@@ -1,6 +1,6 @@
 require ('./components/calculateFilePathTest.js');
 require ('./components/extractFieldsTest.js');
-require ('./components/hasIdAndPropertiesTest.js');
+require ('./components/recordHasIdAndPropertiesTest.js');
 require ('./components/recordHasNameTest.js');
 require ('./components/filterOutUnreadableFilesTest.js');
 require ('./components/isNotDeprecatedRecordTest.js');
